utils/requirements.txt
astroid==3.2.2
attrs==23.2.0
autopep8==2.0.4
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
docstring-to-markdown==0.15
easyocr==1.7.1
evdev==1.7.1
filelock==3.14.0
flake8==7.0.0
fonttools==4.51.0
fsspec==2024.5.0
greenlet==3.0.3
h11==0.14.0
HumanCursor==1.1.5
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
outcome==1.3.0.post0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyGetWindow==0.0.9
pylint==3.2.2
PyMsgBox==1.0.9
pynput==1.7.7
pynvim==0.5.0
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
python-xlib==0.33
python3-xlib==0.15
pytoolconfig==1.3.1
pytweening==1.2.0
PyYAML==6.0.1
requests==2.32.1
rope==1.13.0
scikit-image==0.23.2
scipy==1.13.0
selenium==4.21.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sympy==1.12
tifffile==2024.5.10
tomli==2.0.1
tomlkit==0.12.5
torch==2.3.0
torchvision==0.18.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
whatthepatch==1.0.5
wsproto==1.2.0
yapf==0.40.2
zipp==3.18.2
